requirements.txt
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.1
requests==2.31.0
